Add health monitoring for system app. (#1420)
Add health monitoring for system app / T4R.
The related issues are as follows:

https://github.com/SonySemiconductorSolutions/aitrios-sdk-device-system-apps/issues/1416

I also made the following fix: On Linux, the TerminationReason reason
variable was not initialized, which could result in an undefined value
being returned to the upper layer. To address this, I have added proper
initialization.

Signed-off-by: Seimizu Joukan <Qingshui.Shangguan@sony.com>