Remove Linux reboot workaround for system function (#1452)
This PR removes the workaround code that directly executed reboot
commands on Linux platforms when
`CONFIG_EXTERNAL_SYSTEMAPP_ENABLE_SYSTEM_FUNCTION` was not defined.

- Removed conditional compilation block for Linux reboot workaround
- Deleted direct `execl("/sbin/reboot", "reboot", (char *)NULL)`
execution
- Cleaned up platform-specific reboot handling logic

The workaround was originally implemented to handle reboot functionality
on Linux systems that didn't have system function support. With recent
improvements to the system function infrastructure, this workaround is
no longer necessary and can be safely removed.

Signed-off-by: Seimizu Joukan <Qingshui.Shangguan@sony.com>