Create Temperature threshold config (#1453)
Premise:
- T3P/T3Ws/T5/T4R sensor specifications correspond to T3P side:
Range from -20 to 120.
- RPI has a different specification:
Range from -20 to 80.

Implementation Policy:
- Modify configuration to automatically change ELOG transmission
thresholds according to environment.
- Default value in Kconfig is based on T3P specification (100℃).
- Values that do not require changes remain as default and are not added
to .config.
- Since T3P range goes up to 120, the value 100 lacks a direct basis;
however, testing on the cloud side assumes ELOG triggers at 100℃.
To avoid interface and specification adjustments with the cloud, use the
value 100 as is.
- For RPI, which requires a threshold different from the default, add a
definition setting it to 80 in .config.

Impact Scope:
- Sequence: Affects only abnormal states.
ELOG upload timing changes between 100 and 80.
Since the steady state is around 30 to 50℃, logs should not be
excessively output during normal operation.
- Unit Test (UT): No impact.
UT tests define 100 within internal config files.
There is no change in test logic for either 100 or 80; therefore, simply
adding the definition is sufficient.
- System Test: No impact.
Setting device temperature to 100℃ is not achievable in CI/CD and is
excluded, hence no impact.

Signed-off-by: Seimizu Joukan <Qingshui.Shangguan@sony.com>