Clear EVP Hub URL and port when switching PS to QR. (#1461)
Add process that cleaning the evp hub url and port when switching PS to
QR.

If power off occurs after switching to QR code mode, system transition
to normal mode. That is not as intended.
The reason for switching to normal mode was that conditions were met due
to evp hub url and port remaining. Therefore, in this update, we decided
to remove them.

Preparation: Flash the firmware that includes the changes from this PR,
and write the necessary data to enable booting in normal mode.
1. Execute factory reset. (button pressed over 30 sec)
2. Since the device enters PS mode, press and hold the button for 5 to
29 seconds to switch to QR mode.
3. Scan a QR code configured to connect to a server that returns
is_allowlisted: false.
4. The device enters PS mode after rebooting. When it receives
is_allowlisted: false, it switches to QR mode.
5. Turn off the power while the device is in QR mode.
6. Turn the power back on and confirm that the device enters PS mode.

Signed-off-by: Seimizu Joukan <Qingshui.Shangguan@sony.com>