Add secondary NTP and DNS settings. (#1466)
Extended the functionality to allow configuration of secondary DNS and
NTP settings.

As SIer, I want to set secondary NTP and DNS so that we can prevent from
system down because primary NTP or DNS is down.

The CI build is currently failing. Because we need to update ESF.

- **2nd DNS/NTP settings** are optional. If not included in the QR code,
they will be omitted.
- If a **format error** is detected, it will be handled according to the
QR code error handling procedure.
- If a single space `" "` is set in the QR code, the corresponding 2nd
DNS/NTP setting will be deleted.

Signed-off-by: Seimizu Joukan <Qingshui.Shangguan@sony.com>