Support concurrent streaming and get image. (#1483)
# What changes
Added an option to allow simultaneous operation of streaming and get
image.

# Why
On model EDC, simultaneous operation was allowed by specification, but
the current code did not support it, so we modified it to enable this
functionality.

# Note
When EdgeApp is performing inference, SystemApp can retrieve preview
images,
The AI model follows EdgeApp’s settings and The network_id setting in
DirectGetImage is ignored.
Therefore, skip setting the properties when streaming and DirectGetImage
are executed simultaneously.

Signed-off-by: ryotabando <bando.ryota@miraxia.com>